Add(.auth): separate auth module from user
