+++ src/main/java/com/newbarams/ajaja/module/auth/application/model/AccessToken.java
@@ -1,0 +1,5 @@
+package com.newbarams.ajaja.module.auth.application.model;
+
+public interface AccessToken {
+	String getContent();
+}
