+++ src/main/java/com/newbarams/ajaja/module/auth/application/model/Profile.java
@@ -1,0 +1,11 @@
+package com.newbarams.ajaja.module.auth.application.model;
+
+/**
+ * Get Profile Information only needed to issue Token OR sign-up
+ * @author hejow
+ */
+public interface Profile {
+	Long getOauthId();
+
+	String getEmail();
+}
