+++ src/main/java/com/newbarams/ajaja/module/auth/application/port/in/LoginUseCase.java
@@ -1,0 +1,7 @@
+package com.newbarams.ajaja.module.auth.application.port.in;
+
+import com.newbarams.ajaja.module.auth.dto.AuthResponse;
+
+public interface LoginUseCase {
+	AuthResponse.Token login(String authorizationCode, String redirectUri);
+}
