+++ src/main/java/com/newbarams/ajaja/module/auth/application/service/LoginService.java
@@ -1,18 +1,15 @@
-package com.newbarams.ajaja.module.user.application.service;
+package com.newbarams.ajaja.module.auth.application.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.newbarams.ajaja.global.security.jwt.util.JwtGenerator;
-import com.newbarams.ajaja.module.user.application.model.AccessToken;
-import com.newbarams.ajaja.module.user.application.model.Profile;
-import com.newbarams.ajaja.module.user.application.port.in.LoginUseCase;
-import com.newbarams.ajaja.module.user.application.port.out.AuthorizePort;
+import com.newbarams.ajaja.module.auth.application.model.Profile;
+import com.newbarams.ajaja.module.auth.application.port.in.LoginUseCase;
+import com.newbarams.ajaja.module.auth.application.port.out.AuthorizePort;
+import com.newbarams.ajaja.module.auth.dto.AuthResponse;
 import com.newbarams.ajaja.module.user.application.port.out.CreateUserPort;
 import com.newbarams.ajaja.module.user.application.port.out.FindUserIdByEmailPort;
-import com.newbarams.ajaja.module.user.application.port.out.GetOauthProfilePort;
-import com.newbarams.ajaja.module.user.domain.User;
-import com.newbarams.ajaja.module.user.dto.UserResponse;
 
 import lombok.RequiredArgsConstructor;
 
@@ -21,21 +18,19 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 class LoginService implements LoginUseCase {
 	private final AuthorizePort authorizePort;
-	private final GetOauthProfilePort getOauthProfilePort;
 	private final FindUserIdByEmailPort findUserIdByEmailPort;
 	private final CreateUserPort createUserPort;
 	private final JwtGenerator jwtGenerator;
 
 	@Override
-	public UserResponse.Token login(String authorizationCode, String redirectUri) {
-		AccessToken accessToken = authorizePort.authorize(authorizationCode, redirectUri);
-		Profile profile = getOauthProfilePort.getProfile(accessToken.getContent());
+	public AuthResponse.Token login(String authorizationCode, String redirectUri) {
+		Profile profile = authorizePort.authorize(authorizationCode, redirectUri);
 		Long userId = findIdOrCreateIfNotExists(profile.getEmail(), profile.getOauthId());
 		return jwtGenerator.login(userId);
 	}
 
 	private Long findIdOrCreateIfNotExists(String email, Long oauthId) {
 		return findUserIdByEmailPort.findUserIdByEmail(email)
-				.orElseGet(() -> createUserPort.create(User.init(email, oauthId)));
+				.orElseGet(() -> createUserPort.create(email, oauthId));
 	}
 }
