--- src/main/java/com/newbarams/ajaja/module/user/application/model/AccessToken.java
@@ -1,5 +1,0 @@
-package com.newbarams.ajaja.module.user.application.model;
-
-public interface AccessToken {
-	String getContent();
-}
