--- src/main/java/com/newbarams/ajaja/module/user/application/model/Profile.java
@@ -1,10 +1,0 @@
-package com.newbarams.ajaja.module.user.application.model;
-
-/**
- * Oauth2 Interface
- */
-public interface Profile {
-	Long getOauthId();
-
-	String getEmail();
-}
