--- src/main/java/com/newbarams/ajaja/module/user/application/port/in/ReissueTokenUseCase.java
@@ -1,7 +1,0 @@
-package com.newbarams.ajaja.module.user.application.port.in;
-
-import com.newbarams.ajaja.module.user.dto.UserResponse;
-
-public interface ReissueTokenUseCase {
-	UserResponse.Token reissue(String accessToken, String refreshToken);
-}
