+++ src/main/java/com/newbarams/ajaja/module/user/application/service/ReissueTokenService.java
@@ -4,8 +4,8 @@ import org.springframework.stereotype.Service;
 
 import com.newbarams.ajaja.global.security.jwt.util.JwtGenerator;
 import com.newbarams.ajaja.global.security.jwt.util.JwtValidator;
-import com.newbarams.ajaja.module.user.application.port.in.ReissueTokenUseCase;
-import com.newbarams.ajaja.module.user.dto.UserResponse;
+import com.newbarams.ajaja.module.auth.application.port.in.ReissueTokenUseCase;
+import com.newbarams.ajaja.module.auth.dto.AuthResponse;
 
 import lombok.RequiredArgsConstructor;
 
@@ -16,7 +16,7 @@ class ReissueTokenService implements ReissueTokenUseCase {
 	private final JwtValidator jwtValidator;
 
 	@Override
-	public UserResponse.Token reissue(String accessToken, String refreshToken) {
+	public AuthResponse.Token reissue(String accessToken, String refreshToken) {
 		Long userId = jwtValidator.validateReissuableAndExtractId(accessToken, refreshToken);
 		return jwtGenerator.reissue(userId, refreshToken);
 	}
