+++ src/main/java/com/newbarams/ajaja/module/user/kakao/application/KakaoDisconnectOauthService.java
@@ -3,10 +3,10 @@ package com.newbarams.ajaja.module.user.kakao.application;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
 
-import com.newbarams.ajaja.infra.feign.kakao.KakaoProperties;
-import com.newbarams.ajaja.infra.feign.kakao.KakaoUnlinkFeignClient;
+import com.newbarams.ajaja.infra.feign.kakao.client.KakaoProperties;
+import com.newbarams.ajaja.infra.feign.kakao.client.KakaoUnlinkFeignClient;
+import com.newbarams.ajaja.infra.feign.kakao.model.KakaoUnlinkRequest;
 import com.newbarams.ajaja.module.user.application.port.out.DisconnectOauthPort;
-import com.newbarams.ajaja.module.user.kakao.model.KakaoUnlinkRequest;
 
 import lombok.RequiredArgsConstructor;
