+++ src/test/java/com/newbarams/ajaja/module/user/application/service/UserMockBeans.java
@@ -5,14 +5,13 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 
 @TestConfiguration
 @MockBean({
-	ChangeReceiveTypeService.class,
-	LoginService.class,
-	LogoutService.class,
-	ReissueTokenService.class,
-	RenewNicknameService.class,
-	SendVerificationEmailService.class,
-	VerifyCertificationService.class,
-	WithdrawService.class
+		ChangeReceiveTypeService.class,
+		LogoutService.class,
+		ReissueTokenService.class,
+		RenewNicknameService.class,
+		SendVerificationEmailService.class,
+		VerifyCertificationService.class,
+		WithdrawService.class
 })
 public class UserMockBeans {
 }
